Microphone: Cleanup + renaming
