+++ module_microphone/module.py
@@ -5,13 +5,13 @@ import logging
 # logging.basicConfig(level=logging.DEBUG)
 
 
-def speech_to_text(audio: AudioProcessor):
+def print_speech(audio: AudioProcessor):
     print("Je réfléchis...")
     print("Tu as dis: " + audio.as_text())
 
 
 def scrib():
-    listen.in_background(speech_to_text)
+    listen.for_each_phrase(print_speech)
     print("Le microphone écoute en arrière plan, il écrira ce que tu dis.\n")
     input("Appuies sur une touche pour arrêter.")
     listen.stop()
@@ -23,10 +23,10 @@ def text_to_speech_5_sec():
     print("Tu as dis: " + listen.during("5 secondes").as_text())
 
 
-def record_1_sentence():
+def save_one_phrase():
     """Record one spoken sentence and save it at default path."""
     filepath = (
-        listen.one_sentence().save("/tmp/my_sentence_%Y-%m-%d_%Hh%Mm%Ss.wav").filepath
+        listen.one_phrase().save("/tmp/my_sentence_%Y-%m-%d_%Hh%Mm%Ss.wav").filepath
     )
 
     print("Fichier sauvegardé au chemin " + filepath)
@@ -34,6 +34,6 @@ def record_1_sentence():
 
 def run():
     # Comment / Uncomment functions to try them
-    text_to_speech_5_sec()
-    # scrib()
-    # record_1_sentence()
+    # text_to_speech_5_sec()
+    scrib()
+    # save_one_phrase()
